Global structs are dumped properly
This commit fixes a few bugs in struct_dump(). It gives pointers in structs two bytes of space, and gives arrays in structs the proper amount of space.
A struct_size member is introduced in struct symbol (SYMBOL) to store the size of an array member, and arrays in structs are properly dumped based on this.